chore: fix all clippy warnings and add CHANGELOG.md
Clippy fixes:
- Collapse nested if-let chains into single expressions
- Remove unnecessary borrows and returns
- Rename to_list()/to_list_groups() to into_list()/into_list_groups()
  to follow Rust naming conventions (consuming self)
- Derive Default instead of manual impl for TusksAttr

Add CHANGELOG.md covering v2.1.7, v3.0.0, and v3.1.0.

+++ tusks-test/src/external1.rs
@@ -49,7 +49,7 @@ pub mod tasks {
         println!("  Chain to root: ext1 -> level1 -> root");
         println!("    root_param: {}", params.super_.super_.root_param.as_ref().unwrap());
         
-        Some(sum as u8)
+        Some(sum)
     }
 
     // Nested external module
